Support `column LIKE '%keyword%'`
It's not the original LIKE compatible feature. It's just a short cut of
`column @@ 'keyword'`. So sequential scan result and index scan (or
bitmap scan) results may be different.

If you want to the same result for sequential scan and index scan (or
bitmap scan), use `TokenBigramSplitSymbolAlphaDigit` tokenizer and not
use normalizer.