wal: release a lock every time that a page in a block is initialized
GitHub: fix #150 Reported by BiscuitsColonel Thanks!!!

PostgreSQL occurs PANIC if we truncate PGroonga's WAL that it includes many modifications.
Because the number of Locks that we require for truncating WAL is limited.

The maximum number of these locks is 200.

See:
https://github.com/postgres/postgres/blob/REL_13_1/src/backend/storage/lmgr/lwlock.c#L208

The loop of PGrnWALTruncate() may very well run more than 200 times.
So, we modify that release a lock every running the loop once.