crash-safer: really fix a bug that re-index recovery doesn't work
In this change, crash-safer also tries to run "flush" processes for
all database that uses PGroonga (has "pgrn" file) on start up. Because
crash-safer may need to run "REINDEX INDEX" when Groonga's DB is
broken. It acquires an exclusive lock. We can't do it when the first
"SELECT" that uses a PGroonga index. Because "SELECT" acquires a
shared lock.