Don't define operators into public
GitHub: fix GH-291

It's for backward compatibility for 1.X. but it has a downside. We
can't use PGroonga installed in non-public schema without explicit
`OPERATOR(#{ANOTHER_SCHEMA}.&@)`.

I think that this change doesn't affect to most users because 2.0.0
was released about 6 years ago but this change is a breaking
change. So we will use 3.0.0 for the next version.