wal: update applied position between GenericXLogFinish() and PGrnWALUnlock(data->index)
If we update applied position before GenericXLogFinish(), updated
applied position and written WAL position are inconsistent. Because
written WAL by GenericXLog API are rollback-ed.

This is safe even when this process is killed between
GenericXLogFinish() and PGrnWALUnlock() by the main process (if other
process is crashed, this process is killed by the main process).
Because this process isn't stopped immediately even when SIGTERM is
received. This process is stopped when CHECK_FOR_INTERRUPTS() is
called.