Fix a crash bug in sequential search
It's occurred when:

1. Sequential search with the specified index
   (It means that pgroonga_full_text_search_condition or
   pgroonga_full_text_search_condition_with_scorers is used.)
2. The specified index uses NormalizerTable
3. VACUUM or pgroonga_vacuum() was ran in ANOTHER connection
4. Sequential search with the specified index (This is crashed!)

Sequential search with the specified index is implemented by creating
a internal temporary index. If the specified index uses
NormalizerTable, the internal temporary index refers a persisted
table. But VACUUM or pgroonga_vacuum() was ran in ANOTHER connection,
the persisted table is closed automatically to avoid touching removed
objects. So the internal temporary index refers the closed persisted
table. It causes the crash.

It's fixed by recreating the internal temporary index. It re-refers the
persisted table.