Fix a crash bug in pgroonga_highlight_html()
This is similar to 7bd1e288b88078424fda767312a1fb56e687819c.

It's occurred when:

1. Execute pgroonga_highlight_html() with the specified index
2. The specified index uses NormalizerTable
3. VACUUM or pgroonga_vacuum() was ran in ANOTHER connection
4. Use PGroonga index by SELECT, INSERT and so on to unmap database
5. Execute pgroonga_highlight_html() with the specified index
   (This is crashed!)

Executing pgroonga_highlight_html() with the specified index is
implemented by creating an internal temporary lexicon. If the specified
index uses NormalizerTable, the internal temporary lexicon refers a
persisted table. But VACUUM or pgroonga_vacuum() was ran in ANOTHER
connection, the persisted table is closed automatically to avoid
touching removed objects. So the internal temporary lexicon refers the
closed persisted table. It causes the crash.

It's fixed by recreating an internal highlighter that may have an
internal temporary lexicon. It re-refers the persisted table later.