tools repair-dump: add a script that adjusts pg_dump output
If we use ${table:...} for NormalizerTable and so on, we may not be
able to restore pg_dump output. Because pg_dump output uses empty
search_path. ${table:NO_SCHEMA_INDEX_NAME} doesn't work with empty
search_path. This complement schema information in the given pg_dump
output.

For example, "${table:index}" is converted to "${table:schema.index}".