query-v2: Change to empty when searched with blank space
An error will occur if any of the multiple conditions have a blank space condition.
This is because grn_expr_parse() in PGrnSearchBuildConditionQuery() was not correctly handled
on the PGroonga side when the condition was ignored.

If the condition contains blank spaces, there should be no hits, so the result will be empty.