query-v2: Fix a crash bug with multiple conditions including blank only query condition (#482)
An error will occur if any of the multiple conditions have a blank space
condition.
This is because grn_expr_parse() in PGrnSearchBuildConditionQuery()
doesn't add any condition but PGroonga adds logical AND later. Logical
AND requires 2 conditions but there is only 1 condition in this case. So
the expression becomes invalid.

If the condition contains only blank spaces, there should be no hits, so
the result will be empty.