ci macos: update bundler cache in each PostgreSQL version
GitHub: fix GH-522

Issue

`pgroonga-benchmark` gem couldn't run because of the following error.
Because the `pg_ext.bundle` in `pg` gem is attempting to load the PostgreSQL 14 library (libpq.5.dylib),
while PostgreSQL 15 is installed on the system.

The Cause of this issue

The `setup-ruby` action cached gems.
As a result, the `pg` gem is referencing PostgreSQL 14 because it was cached in a previous workflow with PostgreSQL 14.
Even if the related PostgreSQL version is updated in new workflow, the `pg` gem isn't changed in the cache because
gem isn't updated as it is.

ref: https://github.com/ruby/setup-ruby?tab=readme-ov-file#dealing-with-a-corrupted-cache

How to resolve

By updating the bundler cache for each PostgreSQL version,
this change ensures that the cache does not retain dependencies from different PostgreSQL versions.