wal-resource-manager: improve "CREATE INDEX" and "REINDEX" performance
"CREATE INDEX" and "REINDEX" build an index from existing records. We
can optimize the case.

In general, we create one WAL record for each record. But we can
create one WAL record for all existing records. It improves WAL apply
performance on standby. Because we can avoid "flush" for each record.

This is enabled only when:

* Traditional PGroonga WAL is disabled (`pgroonga.enable_wal = no`) and
* `jsonb` isn't used