wal-resource-manager: add pgroonga.max_bulk_insert_wal_record_size
It's a soft limit of the maximum size of one bulk insert WAL record.

It seems that we can use very large size for one WAL record but it
uses many working memory for generating and reading one bulk insert
WAL record.

The default is 16MiB. We can disable this soft limit by 0.