crash-safe: disable Groonga WAL for source data copy in bulk index creation
Groonga WAL is enabled in WAL resource manager. So this is applied for
not only "pgroonga_crash_safer" but also
"pgroonga_wal_resource_manager" users.

Groonga WAL generation decreases write performance about 1.5-2.0. This
reduces it only for bulk index creation caused by "CREATE INDEX" and
"REINDEX". Other operations such as "INSERT"/"UPDATE"/"DELETE" aren't
affected.

This disables Groonga WAL generation only for the source data copy
phase. The source data copy phase may be failed and it may break
Groonga's source table and its columns. But we don't need to recover
it because the Groonga's source table will be removed in the next
VACUUM. So we can omit Groonga WAL generation for it.