Add support for parallel index build
There are 2 parts in index build:

1. Copy PostgreSQL data to Groonga
2. Construct Groonga index against the copied data

This has implementations of them. But the implementation of 1. is
disabled by default. We can enable by `SET
pgroonga.enable_parallel_build_copy = true` but we don't document
it. Because it's not useful for now.

I thought that we can improve performance of 1. by parallel
workers. But it was wrong. It just adds overhead. It doesn't improve
performance. It just slowed down. So the current implementation may
work but it's unstable. I didn't test carefully.

We may revisit parallel workers but we disable it for now.

The implementation of 2. (`grn_ctx_set_n_workers(ctx,
indexInfo->ii_ParallelWorkers + 1)`) is enabled by default. Because it
improves performance in many cases. We may want to use more threads
than `ii_ParallelWorkers` for performance but we use
`ii_ParallelWorkers + 1` for now.

This is tested in Groonga. So this is stable.