ci: Fix incus device add error
With the latest GitHub Actions image, we get the following error:

```
Error: Failed to start device "host-rw": Required idmapping abilities not available
```

Same issue as this one:
https://discuss.linuxcontainers.org/t/error-failed-to-setup-device-mount-idmapping-abilities-are-required-but-arent-supported-on-system/19976

If the `shift=true` option is removed, this error will not occur.
However, when that option is removed, the GH-721 issue reoccurs.

That permission error could be avoided by creating the directory with mode=777.
So, remove the `shift=true` option and create the directory with mode=777.