custom-scan: Add PGrnChooseIndex()
This commit is part of the work to implement a custom scan.

Custom scan needs to implement logic to choose which indexes to use.
Because our first goal is to be able to return records, we choose the first PGroonga index
we find in the for loop in a temporary implementation.