custom-scan: Fetch and return data from Groonga resources
This commit is part of the work to implement a custom scan.

The first goal is to get the results of the following query as expected.
```sql
/* sql/custom-scan/debug.sql */
 SELECT content
   FROM memos
  WHERE content &@~ 'PGroonga OR Groonga';
```

This commit does not consider the conditions of the `WHERE` clause and fetches all data from Groonga's sources table.
The goal of the next PR is to filter by the conditions of the `WHERE` clause.