Add support for `column &@ IN (keyword1, keyword2, ...)` (#716)
PostgreSQL 18 optimizes `column &@ keyword1 OR column &@ keyword2 OR
...` to `column &@ IN (keyword1, keyword2, ...)` / `column &@ ANY
('{keyword1,keyword2,...}'::text[])`.

We need to add support `&@` + `IN`/`ANY` for PostgreSQL 18 support.

Note that `column &@ keyword1 OR column &@ keyword2 OR ...` with index
scan doesn't use index scan with PostgreSQL 17 or older. (Bitmap scan
can be used.) We don't add support for index scan for `column &@
keyword1 OR column &@ keyword2 OR ...` with PostgreSQL 17 or older.

We should remove `_{1,2}.out` expected files when we drop support for
PostgreSQL 17 or older. They are only for old PostgreSQL.

---------

Co-authored-by: Sutou Kouhei <kou@clear-code.com>