custom-scan: Prepare to support parallel scan
Added the necessary functions. First, add them as empty:
* PGrnEstimateDSMCustomScan
* PGrnInitializeDSMCustomScan
* PGrnReInitializeDSMCustomScan
* PGrnInitializeWorkerCustomScan

We also added a NodeTag at the head of PGrnScanIndexData.
PGrnScanIndexData will be added to custom_private, but the elements added to custom_private must be Node.
Without a NodeTag, the following error occurs when running in parallel.

```
ERROR:  unrecognized node type: XXX
```