custom-scan: Set the value of `custom_private` to `Node` only (#830)
We need to set only `Node` to `custom_private`, especially when using
parallel scan.
In the code before the modification, `PGrnScanIndexData` and
`ScanKeyData` are not `Node`, so an error occurs during parallel scan.

Terefore, we replaced them with `Node` type provided by PostgreSQL, such
as `List`.

Although it is possible to make `PGrnScanIndexData` `Node`, it is
simpler to use the types provided by PostgreSQL.