test term-search/varchar-array/contain-v2: fix row level security tests
GitHub: GH-849

The first test row used user_name 'nonexistent' with content 'PostgreSQL'.
Given the query `tags &> 'Groonga'`, the RLS check was used but doesn't effect the last result.
It's because 'PostgreSQL' doesn't match with 'Groonga'.

If we add the additional record with "ARRAY['PostgreSQL', 'Groonga']" that matches with 'Groonga', we can confirm that the row is correctly filtered by the RLS policy.

Note: Similar fixes will follow separately.