test term-search/varchar/equal/row-level-security: fix row level security tests
GitHub: GH-849

The first test row used `nonexistent` as `user_name` with content '`PostgreSQL`'. Given the query `name &= 'groonga'`, the RLS check was used but doesn't effect the last result. It's because `'PostgreSQL'` doesn't contain `'groonga'`.

If we add the additional record with `'GROONGA'` that contains `'groonga'`, we can confirm that the row is correctly filtered by the RLS policy.

Note: Similar fixes will follow separately.