test full-text-search/text/single/match-condition-v2: fix row level security tests
GitHub: GH-849

The first test row used `nonexistent` as `user_name` with content `'PostgreSQL is a RDBMS.'`.
Given the query `''content &@ pgroonga_condition('Groonga', index_name => 'pgrn_index')`,
the RLS check is used but does not effect the last result.
It's because `'PostgreSQL is a RDBMS.'` doesn't match the above query's condition.

If we add an additional record with `'Rroonga is the Ruby bindings that uses Groonga.'` that matches the above query's condition,
we can confirm that the row is correctly filtered by the RLS policy.

Note: Similar fixes will follow separately.