test full-text-search/text/single/similar-v2: fix row level security tests (#893)
GitHub: GH-849

The first test row used `nonexistent` as `user_name` with content
`'PostgreSQL is an OSS RDBMS'`.

Given the query `"content &@* 'Mroonga is a MySQL plugin that uses
Groonga.'"`, the RLS check is used but does not effect the last result.
It's because the first record's content doesn't match the above query's
condition.

If we add an additional record with `'Rroonga is the Ruby binding that
uses Groonga.'` that matches the above query's condition, we can confirm
that the row is correctly filtered by the RLS policy.

Note: Similar fixes will follow separately.