Add a  similar distance operator (#897)
This works only with index scan. `<&@*>` always returns `0.0` but it's
not used in index scan based sort.

If sequential scan or bitmap scan is used, sort doesn't work. Parallel
scan may not work too. (It's not tested.)

We want to use `LIMIT` value to reduce sort targets but we can't use
`LIMIT` value in index scan...

---------

Co-authored-by: Sutou Kouhei <kou@clear-code.com>