test regexp/varchar/regexp-v2: fix row level security tests (#904)
GitHub: GH-849

The first test row used `nonexistent` as `user_name` with content
`'PostgreSQL is a RDBMS.'`.

Given the query `content &~ 'groonga'`, the RLS check is used but does
not effect the last result. It's because the first record's content
doesn't match the above query's condition.

If we add an additional record with `'Rroonga is the Ruby bindings of
Groonga.'` that matches the above query's condition, we can confirm that
the row is correctly filtered by the RLS policy.

Note: Similar fixes will follow separately.