semantic-search: add support for large (4KiB+) embedding
This require Groonga 15.2.1 or later.

If the specified language model uses 1025+ dimensions for embedding,
we can't store it as a Groonga's table key. Because 1025+ dimensions
embedding has 4KiB+ size and Groonga's table key's max size is 4KiB.

In the case, we store embedding to column not table key.

This also adds the `n_gpu_layers` option. We can use it for disabling
GPU. If the specified language model is large, we may not be able to
load it to (small) GPU. In the case, we can disable GPU. (It works but
slow...)